fix: Redesign a method of authorization by creating a component that returns data in bytes array
