+++ src/main/java/com/nerdtranslator/lingueeapibridge/service/CredentialsProvider.java
@@ -1,0 +1,7 @@
+package com.nerdtranslator.lingueeapibridge.service;
+
+import org.springframework.stereotype.Component;
+
+public interface CredentialsProvider {
+    byte[] getAuthorizationData();
+}
