+++ src/main/java/com/nerdtranslator/lingueeapibridge/service/NaturalLangApiService.java
@@ -1,0 +1,5 @@
+package com.nerdtranslator.lingueeapibridge.service;
+
+public interface NaturalLangApiService {
+    String analyzeText(String text);
+}
