--- src/main/java/com/nerdtranslator/lingueeapibridge/service/NaturalLangService.java
@@ -1,5 +1,0 @@
-package com.nerdtranslator.lingueeapibridge.service;
-
-public interface NaturalLangService {
-    String analyzeTextUsingApi(String text);
-}
