+++ src/main/java/com/nerdtranslator/lingueeapibridge/service/TextToSpeechApiService.java
@@ -1,0 +1,5 @@
+package com.nerdtranslator.lingueeapibridge.service;
+
+public interface TextToSpeechApiService {
+    byte[] transformTextToSound(String textToTransfer, String langCode);
+}
