--- src/main/java/com/nerdtranslator/lingueeapibridge/service/TextToSpeechService.java
@@ -1,5 +1,0 @@
-package com.nerdtranslator.lingueeapibridge.service;
-
-public interface TextToSpeechService {
-    byte[] transformTextToSoundUsingApi(String textToTransfer, String langCode);
-}
